update license information
We are adapting splicePump and spliceDrain from the stdlib. We also
plan to import the standard library splice tests, from package net.
Distribute package zerocopy under the BSD license, to avoid the mix
of ISC and BSD licenses.

Add The Go Authors to copyright headers where we use or adapt code
from the Go standard library.

Signed-off-by: Andrei Tudor Călin <mail@acln.ro>